REFACTOR: Update AuthorCompactProfile to functional component and use TypeScript
